refactor(server): scope ConfigKV to the resource API
