+++ server/apps/api/package.json
@@ -44,7 +44,6 @@
     "@opentelemetry/sdk-trace-node": "catalog:",
     "@opentelemetry/semantic-conventions": "catalog:",
     "@proj-airi/auth-shared": "workspace:*",
-    "@proj-airi/config-shared": "workspace:*",
     "@proj-airi/drizzle-migration": "workspace:*",
     "@proj-airi/drizzle-orm-browser-migrator": "catalog:",
     "@proj-airi/server-sdk-shared": "workspace:*",
