+++ server/apps/api/src/schemas/index.ts
@@ -1,5 +1,6 @@
 export * from './characters'
 export * from './chats'
+export * from './config-kv'
 export * from './flux'
 export * from './flux-transaction'
 export * from './llm-request-log'
@@ -10,4 +11,3 @@ export * from './stripe'
 export * from './user-character'
 export * from './voice-packs'
 export * from '@proj-airi/auth-shared'
-export * from '@proj-airi/config-shared'
