--- server/packages/config-shared/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@proj-airi/config-shared",
-  "type": "module",
-  "version": "0.11.3",
-  "private": true,
-  "description": "Shared ConfigKV persistence and invalidation contracts for AIRI services",
-  "exports": "./src/index.ts",
-  "scripts": {
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "drizzle-orm": "catalog:",
-    "ioredis": "catalog:"
-  }
-}
