--- server/packages/config-shared/src/index.ts
@@ -1,3 +1,0 @@
-export * from './contracts'
-export * from './schema'
-export * from './store'
